git-editfile.sh: Check to ensure the file existed before the backup.

+++ scripts/helper-git/git-editfile/git-editfile.sh
@@ -24,6 +24,7 @@ shift $((OPTIND - 1))
 
 backup() {
     local file=$1
+    [ ! -f $file ] && return 0
     cp $file $file.bak
 }
